+++ src/threads.js
@@ -4753,7 +4753,7 @@ Process.prototype.goToLayer = function (name) {
 
 // Process scene primitives
 
-Process.prototype.doSwitchToScene = function (id, transmission) {
+Process.prototype.doSwitchToScene = function (id, transmission) { // +++
     var rcvr = this.blockReceiver(),
         idx = 0,
         message = this.inputOption(transmission.at(1)),